+++ src/Providers/DfServiceProvider.php
@@ -46,6 +46,7 @@ class DfServiceProvider extends ServiceProvider
                 'ADLdap',
                 'Aws',
                 'Azure',
+                'AzureAD',
                 'Cache',
                 'Cassandra',
                 'Couchbase',